Updated MPI mesh
Previously, the MPI mesh was made in 1D so the meshgrid was not being appropriately split by MPI to be as efficient as possible. I have changed the meshgrid in the domain variable to have a 2D mesh which is possible in 3D since there are 2 fourier basis and one chebyshev. Currently it is set to 4x4 so that with a 16 core MPI parralelisation, each core will be solving 1/4 of the grid. The dealias was also changed to 2 so that there is never an odd numver of gridpoints during calculations. as this would make it impossible to split the grid evenly